+++ packages/cli/src/fingerprint/fingerprintWatchCommand.ts
@@ -10,6 +10,7 @@ import { verbose } from '../utils';
 import { FingerprintEvent } from './fingerprinter';
 import FingerprintQueue from './fingerprintQueue';
 import Globber from './globber';
+import path from 'path';
 
 export default class FingerprintWatchCommand {
   private pidfilePath: string | undefined;
@@ -17,6 +18,8 @@ export default class FingerprintWatchCommand {
   public watcher?: FSWatcher;
   private poller?: Globber;
   private _numProcessed = 0;
+  public unreadableFiles = new Set<string>();
+  public symlinkLoopFiles = new Set<string>();
 
   public get numProcessed() {
     return this._numProcessed;
@@ -45,25 +48,126 @@ export default class FingerprintWatchCommand {
     }
   }
 
-  async watcherErrorFunction (error: Error) {
-    if (this.watcher && error.message.includes("ENOSPC: System limit for number of file watchers reached")) {
-      console.warn(error.stack);
-      console.warn("Will disable file watching. File polling will stay enabled.");
+  async disableFileWatching() {
+    if (this.watcher) {
+      console.warn('Will disable file watching. File polling will stay enabled.');
       await this.watcher?.close();
       this.watcher = undefined;
-      console.warn("File watching disabled.");
+      console.warn('File watching disabled.');
+    }
+  }
+
+  getFilenameFromErrorMessage(errorMessage: string): string {
+    // The errorMessage must contain the filename within single quotes
+    // and looks like this:
+    // Error: UNKNOWN: unknown error, lstat 'c:\Users\Test\Programming\MyProject'
+    let quoteStartIndex = errorMessage.indexOf(`'`);
+    let quoteEndIndex = errorMessage.indexOf(`'`, quoteStartIndex + 1);
+    return errorMessage.substring(quoteStartIndex + 1, quoteEndIndex);
+  }
+
+  isError(error: unknown, code: string): boolean {
+    const err = error as NodeJS.ErrnoException;
+    return err.code === code;
+  }
+
+  dontProcessFileAgain(error: Error, filename: string, telemetryName: string, files: Set<string>) {
+    console.warn(error.stack);
+    Telemetry.sendEvent({
+      name: telemetryName,
+      properties: {
+        errorMessage: error.message,
+        errorStack: error.stack,
+      },
+    });
+    files.add(filename);
+    console.warn('Will not read the file ' + filename + ' again.');
+  }
+
+  getToplevelSymbolicLink(filename: string): string {
+    // given a filename like '/tmp/topdir/dir1/point_up/dir1/point_up/dir1/point_up'
+    // lastToken is set to 'point_up' (the actual symlink)
+    // directoryPostfix is set to '/dir1/point_up'
+    // topLevelSymbolicLink is set to '/tmp/topdir/dir1/point_up'
+    const splitTokens = filename.split(path.sep);
+    const lastToken = splitTokens[splitTokens.length - 1];
+    const lastIndex = filename.lastIndexOf(lastToken);
+    const secondToLastIndex = filename.lastIndexOf(lastToken, lastIndex - 1);
+    const directoryPostfixLength = lastIndex - secondToLastIndex;
+    const directoryPostfix = filename.substring(
+      filename.length - directoryPostfixLength,
+      filename.length
+    );
+    const splitDirectoryPostfix = filename.split(directoryPostfix);
+    const toplevelSymbolicLink = splitDirectoryPostfix[0] + directoryPostfix;
+
+    return toplevelSymbolicLink;
+  }
+
+  async watcherErrorFunction(error: Error) {
+    if (this.isError(error, 'ENOSPC')) {
+      console.warn(error.stack);
+      await this.disableFileWatching();
       Telemetry.sendEvent({
-          name: `index:watcher_error:enospc`,
-          properties: {
-            errorMessage: error.message,
-            errorStack: error.stack,
-          },
-        });
+        name: `index:watcher_error:enospc`,
+        properties: {
+          errorMessage: error.message,
+          errorStack: error.stack,
+        },
+      });
+    } else if (this.isError(error, 'EMFILE')) {
+      // Don't crash if too many files are open. When the files close
+      // the indexer will pick them up.
+      console.warn(error.stack);
+      await this.disableFileWatching();
+      Telemetry.sendEvent({
+        name: `index:watcher_error:emfile`,
+        properties: {
+          errorMessage: error.message,
+          errorStack: error.stack,
+        },
+      });
+    } else if (this.isError(error, 'UNKNOWN') && error.message.includes('lstat')) {
+      const filename = this.getFilenameFromErrorMessage(error.message);
+      this.dontProcessFileAgain(
+        error,
+        filename,
+        `index:watcher_error:unknown`,
+        this.unreadableFiles
+      );
+    } else if (this.isError(error, 'ELOOP')) {
+      const filename = this.getFilenameFromErrorMessage(error.message);
+      const filenameTopLevelSymbolicLink = this.getToplevelSymbolicLink(filename);
+      this.dontProcessFileAgain(
+        error,
+        filenameTopLevelSymbolicLink,
+        `index:watcher_error:eloop`,
+        this.symlinkLoopFiles
+      );
     } else {
       // let it crash if it's some other error, to learn what the error is
       throw error;
     }
-  };
+  }
+
+  // Custom ignore function needed to skip unreadableFiles because
+  // attempting to read them can block reading all files. It also
+  // cuts down the watch tree to just what we need.
+  ignored(targetPath: string) {
+    ['/node_modules/', '/.git/'].forEach((pattern) => {
+      if (targetPath.includes(pattern)) {
+        return true;
+      }
+    });
+
+    if (this.unreadableFiles.has(targetPath) || this.symlinkLoopFiles.has(targetPath)) {
+      return true;
+    }
+
+    // Also make sure to not try to recurse down node_modules or .git
+    const basename = path.basename(targetPath);
+    return basename === 'node_modules' || basename === '.git';
+  }
 
   async execute() {
     this.fpQueue.process().then(() => {
@@ -83,7 +187,7 @@ export default class FingerprintWatchCommand {
 
     this.watcher = watch(glob, {
       ignoreInitial: true,
-      ignored: ['**/node_modules/**', '**/.git/**'],
+      ignored: this.ignored.bind(this),
       ignorePermissionErrors: true,
     })
       .on('add', this.added.bind(this))